Refactoring Redux Store code
