--- src/redux/actionTypes.js
@@ -1,2 +1,0 @@
-export const ADD_BOOK = 'addBook';
-export const REMOVE_BOOK = 'removeBook';
